requirements.txt
click
PyGithub
ruamel.yaml
pyyaml